feat(sentinel): <- use configurable gas prices/limits in that

+++ v3_bridges/sentinel/src/lib/config/config_traits.rs
@@ -2,7 +2,7 @@ use common::BridgeSide;
 use ethereum_types::Address as EthAddress;
 
 pub trait ConfigT {
-    fn gas_limit(&self) -> u64;
+    fn gas_limit(&self) -> usize;
     fn side(&self) -> BridgeSide;
     fn is_validating(&self) -> bool;
     fn router(&self) -> EthAddress;
